+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the electron-react-app (ERA) are listed here.
 - Create about page
 - Update layout to new design
 - Self host auth0 callback page [(#58)](https://github.com/heyito/ito/pull/58)
+- Better handle errors from transcription service [(#62)](https://github.com/heyito/ito/pull/62)
 
 ### v0.2.2